requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
google-genai==0.1.0
